Changed Integer to Double when calling toString() on FRAME_RATE_CAP

+++ trunk/workflow-manager/src/main/java/org/mitre/mpf/wfm/util/AggregateJobPropertiesUtil.java
@@ -460,7 +460,7 @@ public class AggregateJobPropertiesUtil {
 
         if (frameRateCapPropInfo.getLevel() == PropertyLevel.NONE) {
             frameRateCapPropInfo = new PropertyInfo(MpfConstants.FRAME_RATE_CAP_PROPERTY,
-                                                    Integer.toString(systemFrameRateCap), PropertyLevel.SYSTEM);
+                                                    Double.toString(systemFrameRateCap), PropertyLevel.SYSTEM);
         }
 
         PropertyInfo propInfoToUse;
